requirements.txt
python-dotenv
vk_api
flask
